requirements.txt
asn1crypto
certifi
cffi
charset-normalizer
cryptography
filelock
idna
numpy
oscrypto
packaging
pandas
pycparser
pycryptodomex
PyJWT
pyOpenSSL
python-dateutil
python-dotenv
pytz
requests
six
snowflake-connector-python
typing_extensions
tzdata
urllib3
